requirements.txt
playwright
pytest-playwright
pyyaml